app/requirements.txt
fastapi[all]
PyJWT
passlib
uvicorn
pydantic
psycopg2-binary
sqlalchemy
asyncpg
databases
alembic
greenlet
isbnlib
